+++ app-core/src/main/kotlin/org/archivekeep/app/core/persistence/drivers/filesystem/FileSystemStorageDriver.kt
@@ -2,27 +2,31 @@ package org.archivekeep.app.core.persistence.drivers.filesystem
 
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.FlowCollector
 import kotlinx.coroutines.flow.distinctUntilChanged
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onEach
-import kotlinx.coroutines.flow.transform
 import org.archivekeep.app.core.domain.repositories.RepoAuthRequest
 import org.archivekeep.app.core.domain.storages.Storage
 import org.archivekeep.app.core.domain.storages.StorageConnection
 import org.archivekeep.app.core.domain.storages.StorageDriver
 import org.archivekeep.app.core.domain.storages.StorageInformation
 import org.archivekeep.app.core.utils.ProtectedLoadableResource
-import org.archivekeep.app.core.utils.exceptions.DisconnectedStorageException
+import org.archivekeep.app.core.utils.generics.OptionalLoadable
+import org.archivekeep.app.core.utils.generics.OptionalLoadable.LoadedAvailable
+import org.archivekeep.app.core.utils.generics.OptionalLoadable.NotAvailable
 import org.archivekeep.app.core.utils.generics.UniqueSharedFlowInstanceManager
+import org.archivekeep.app.core.utils.generics.mapLoadedData
+import org.archivekeep.app.core.utils.generics.mapLoadedDataAsOptional
 import org.archivekeep.app.core.utils.generics.mapLoadedDataToOptional
+import org.archivekeep.app.core.utils.generics.mapToOptionalLoadable
 import org.archivekeep.app.core.utils.generics.sharedGlobalWhileSubscribed
 import org.archivekeep.app.core.utils.identifiers.RepositoryURI
 import org.archivekeep.app.core.utils.identifiers.StorageURI
 import org.archivekeep.files.repo.Repo
 import org.archivekeep.files.repo.files.openFilesRepoOrNull
-import org.archivekeep.utils.loading.Loadable
 import org.archivekeep.utils.loading.mapLoadedData
 import org.archivekeep.utils.loading.waitLoadedValue
+import java.nio.file.Path
 import kotlin.io.path.Path
 
 class FileSystemStorageDriver(
@@ -54,7 +58,7 @@ class FileSystemStorageDriver(
         StorageConnection(
             storageURI,
             fileStores.mountedFileSystems
-                .mapLoadedDataToOptional { connectedFSList ->
+                .mapLoadedDataAsOptional { connectedFSList ->
                     connectedFSList
                         .filter { it.fsUUID == storageURI.data }
                         .firstOrNull()
@@ -75,70 +79,51 @@ class FileSystemStorageDriver(
 
         val uriDATA = uri.typedRepoURIData as FileSystemRepositoryURIData
 
-        return fileStores
-            .mountedFileSystems
-            .transform {
+        return getPathInFileSystem(uriDATA)
+            .distinctUntilChanged()
+            .map {
                 when (it) {
-                    is Loadable.Failed -> throw it.throwable
-                    is Loadable.Loaded -> emit(it.value.firstOrNull { it.fsUUID == uriDATA.fsUUID })
-                    Loadable.Loading -> {}
+                    OptionalLoadable.Loading -> ProtectedLoadableResource.Loading
+                    is OptionalLoadable.Failed -> ProtectedLoadableResource.Failed(it.cause)
+                    is NotAvailable -> ProtectedLoadableResource.Failed(it.cause ?: RuntimeException("Not found path in file system"))
+                    is LoadedAvailable -> {
+                        println("Open $uriDATA")
+
+                        val openedRepo = openFilesRepoOrNull(it.value)
+
+                        if (openedRepo == null) {
+                            ProtectedLoadableResource.Failed(RuntimeException("Not repo"))
+                        } else {
+                            ProtectedLoadableResource.Loaded(openedRepo)
+                        }
+                    }
                 }
-            }.distinctUntilChanged()
-            .transform { connectedFS ->
-                accessor(uriDATA, connectedFS)
             }.onEach {
                 println("New repo accessor for $uriDATA: $it")
             }
     }
 
-    private suspend fun FlowCollector<ProtectedLoadableResource<Repo, RepoAuthRequest>>.accessor(
-        repo: FileSystemRepositoryURIData,
-        connectedFS: MountedFileSystem?,
-    ) {
+    fun getPathInFileSystem(repo: FileSystemRepositoryURIData): Flow<OptionalLoadable<Path>> {
         val pathInFS = repo.pathInFS
 
-        println("Open $repo in $connectedFS")
-
-        if (connectedFS == null) {
-            println("ERROR: FS not found: ${repo.fsUUID}")
-            emit(
-                ProtectedLoadableResource.Failed(
-                    object :
-                        RuntimeException("ERROR: FS not found: ${repo.fsUUID}"),
-                        DisconnectedStorageException {},
-                ),
-            )
-            return
-        }
-
-        val mp =
-            connectedFS.mountPoints
-                .filter { pathInFS.startsWith(it.fsSubPath) }
-                .maxByOrNull { it.fsSubPath.length }
-
-        if (mp == null) {
-            println("ERROR: mount point not found: $pathInFS in ${connectedFS.mountPoints}")
-            emit(
-                ProtectedLoadableResource.Failed(
-                    object :
-                        RuntimeException("ERROR: mount point not found: $pathInFS in ${connectedFS.mountPoints}"),
-                        DisconnectedStorageException {},
-                ),
-            )
-            return
-        }
-
-        val p =
-            Path(mp.mountPath).resolve(
-                pathInFS.removePrefix(mp.fsSubPath).removePrefix("/"),
-            )
-
-        val openedRepo = openFilesRepoOrNull(p)
-
-        if (openedRepo == null) {
-            emit(ProtectedLoadableResource.Failed(RuntimeException("Not repo")))
-        } else {
-            emit((ProtectedLoadableResource.Loaded(openedRepo)))
-        }
+        return fileStores
+            .mountedFileSystems
+            .mapToOptionalLoadable()
+            .mapLoadedDataToOptional { mountedFileSystems ->
+                mountedFileSystems
+                    .firstOrNull { it.fsUUID == repo.fsUUID }
+                    ?.let { LoadedAvailable(it) }
+                    ?: NotAvailable(FileSystemNotFoundException(repo.fsUUID))
+            }.mapLoadedDataToOptional { connectedFS ->
+                connectedFS.mountPoints
+                    .filter { pathInFS.startsWith(it.fsSubPath) }
+                    .maxByOrNull { it.fsSubPath.length }
+                    ?.let { LoadedAvailable(it) }
+                    ?: NotAvailable(MountPointNotFoundForPathException(pathInFS, connectedFS))
+            }.mapLoadedData { mp ->
+                Path(mp.mountPath).resolve(
+                    pathInFS.removePrefix(mp.fsSubPath).removePrefix("/"),
+                )
+            }
     }
 }