+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - feature: relocations with duplication decrease support
+- UI: open repository folder in system's file manager application
 - UI: support relocations in Upload/Download dialogs (including duplication increase and/or decrease)
 - UI: Material Design TopAppBar (with client side decorations)
 - build: DEB and RPM packaging support